moved all services utilizing internal database to dbService
